requirements_inference.txt
torch
torchvision
torchaudio
facenet-pytorch
numpy
Pillow